Persist offline changes incrementally instead of re-snapshotting
Port of yorkie-js-sdk#1354, completing the interface landed in 7ee8c021bd.

`Client` now appends each local change to the store's log and only writes a
full snapshot when `shouldCompact` says the log has outgrown the snapshot it
is appended to. Re-serializing the whole document per edit cost time
proportional to the document, and a document is at its largest while it is
being edited.

Restore is snapshot, then header, then log. The header has to be applied
before the replay: it carries the checkpoint the last sync acked, and the
replay needs it to know which appended changes the server already has. Those
are still applied -- the log is the delta between the snapshot and the
current content, so skipping one would leave the root behind -- but they are
not re-queued for push.

A sync now records the header alone rather than re-persisting. An online
client syncs constantly, and re-snapshotting per sync is the cost this
release exists to remove. The log is deliberately left untrimmed: it is also
the delta to current content, and nothing brings the snapshot forward on a
push-ack, so dropping acked entries would leave that content in neither
place while the persisted `serverSeq` claims the server has it.

A failed append poisons the log rather than continuing into a hole, and the
next local change repairs it with a fresh snapshot -- which loses nothing,
since a snapshot embeds the whole pending queue, while appending into a
holed log would lose the entire offline session at restore.

`Document` gains `metaToBytes` / `restoreMetaFromBytes` for the header, and
`getPendingChangesAfter` / `restoreAppendedChanges` for the log. The stored
change is a serialized `PbChange` rather than upstream's JSON `ChangeStruct`,
matching how `toBytes` already encodes pending changes on iOS.

`ErrDocumentOpenElsewhere` replaces the generic argument error when the
single-active-session lock is held, so a consumer can fall back to a
non-persisting client without matching on message text.

The base write reads the snapshot and the carried pending changes at the
same instant. Reading the latter after awaiting the write let a change
minted during it count as already embedded in a snapshot taken before it,
which meant it was never appended -- losing that edit outright. Caught by
`test_chained_persists_land_in_order_even_when_the_first_write_is_slower`,
now rewritten against the append model.

+++ Sources/Core/DocStore.swift
@@ -22,11 +22,11 @@ import Foundation
 /// what lets a restore detect a hole rather than replaying a discontinuous run.
 public struct StoredChange: Sendable, Equatable {
     /// The `clientSeq` this change carries.
-    public let clientSeq: Int64
+    public let clientSeq: UInt32
     /// The serialized change.
     public let bytes: Data
 
-    public init(clientSeq: Int64, bytes: Data) {
+    public init(clientSeq: UInt32, bytes: Data) {
         self.clientSeq = clientSeq
         self.bytes = bytes
     }
